+++ junos/provider.go
@@ -208,6 +208,7 @@ func Provider() *schema.Provider {
 			"junos_system_syslog_file":                                   resourceSystemSyslogFile(),
 			"junos_system_syslog_host":                                   resourceSystemSyslogHost(),
 			"junos_vlan":                                                 resourceVlan(),
+			"junos_vstp":                                                 resourceVstp(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"junos_interface":                   dataSourceInterface(),