+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ENHANCEMENTS:
 * add `junos_lldpmed_interface` resource (Fixes parts of #340)
 * add `junos_rstp` resource (Fixes parts of #340)
 * add `junos_rstp_interface` resource (Fixes parts of #340)
+* add `junos_vstp` resource
 
 BUG FIXES: